This would add tests to ensure that the machine-readable lists are sorted.
I'm not actually saying they should be; right now, anyway, they are not.
ucd-list.txt would take a 200-line diff to become sorted.
ucd-list-deprecated.txt currently has "comments" in it and is sorted by
deprecation date.

This of course raises the question: Does it buy us anything to have these
files sorted?  Sure: it wouldn't be hard to do, but perhaps it's
actually preferable to have the Johnson bands in the conventional sequence
U, B, V, R, I, J, H, K?

I have also put in the auto-pulling of ivoatex from the current ivoatex
Makefile.  If this doesn't get merged because, that part should at least
go in.